test: Rename `parse\big-file-performance` test suite
